Report the havoced values into the ktest files.
This upgrade requires the client to declare all memory location that may
get havoced before antering the analyzed loop.

Points to imporove:
- Error reporting. Right now klee crashes as soon as it tries to havoc
the value that was not declared by the client beforehand as amenable to
havocing.
- Analysis of multiple loops. Currently only the havoc value per mem
location is global, and gets rewritten at every loop that is havoced.
Need to associate the values with the loops (i.e. add an unique ID
argument to the klee_induce_invariants)

Add feature tests.
Make dumping havoced values opt-in, to get the backwards compatibility.
Make declaration of the locations that may get havoced optional.